Got rid of those ugly switch statements and long html variables
Instead of storing each bit of markup for each workout in its own variable, I created an array of objects containing the workout data.
Every time the next or previous button is clicked, it increments/decrements x then loops through the array of objects and sets each objects value to a variable. If x equals the "value" value, it pulls the data from that object and interpolates the variable into a single html template. It always keeps track of where the user is without directly checking using switch or if/else statements.